+++ mRemoteNG/Properties/AssemblyInfo.cs
@@ -10,7 +10,8 @@ using System.Resources;
 
 // Compute version values
 
-//Build nr: 3621
+
+//Fork release: 2
 
 // General Information
 [assembly: AssemblyTitle("mRemoteNG")]
@@ -23,7 +24,7 @@ using System.Resources;
 [assembly: AssemblyCulture("")]
 
 // Version information
-[assembly: AssemblyVersion("1.78.2.3621")]
-[assembly: AssemblyFileVersion("1.78.2.3621")]
+[assembly: AssemblyVersion("1.78.2.2")]
+[assembly: AssemblyFileVersion("1.78.2.2")]
 [assembly: NeutralResourcesLanguageAttribute("en-US")]
-[assembly: AssemblyInformationalVersion("1.78.2 (Nightly Build 3621) x64")]
+[assembly: AssemblyInformationalVersion("1.78.2-yll.2 x64")]